Splitted code on sections
